+++ packages/nft-checkout/src/operations/evm/helpers/chain.ts
@@ -1,6 +1,7 @@
-import { ChainNames } from '../../../enums'
-import { CHAIN_IDS } from '../../../const'
 import { ChainTypes } from '@rarimo/provider'
+
+import { CHAIN_IDS } from '@/const'
+import { ChainNames } from '@/enums'
 
 const chains = CHAIN_IDS[ChainTypes.EVM]