+++ packages/nft-checkout/src/operations/evm/helpers/estimator/slippage.ts
@@ -1,4 +1,4 @@
-import { errors } from '../../../../errors'
+import { errors } from '@/errors'
 
 export const validateSlippage = (slippage: number) => {
   if (slippage < 0 || slippage > 100) {