+++ packages/nft-checkout/src/operations/evm/helpers/index.ts
@@ -1,4 +1,4 @@
 export * from './chain'
-export * from './tokens'
 export * from './estimator/'
 export * from './payment'
+export * from './tokens'