+++ packages/provider/src/enums/index.ts
@@ -1,3 +1,4 @@
-export * from './providers'
 export * from './chains'
+export * from './near-wallet-action-type'
+export * from './providers'
 export * from './solana-rpc-error-codes'