+++ packages/provider/src/types/provider.ts
@@ -5,19 +5,22 @@ import {
   TransactionSignature,
 } from '@solana/web3.js'
 import { ethers, providers } from 'ethers'
-
-import { ChainTypes, Providers } from '../enums'
-import { EthereumProvider } from './ethereum'
+import { providers as nearProviders } from 'near-api-js'
 import { TransactionConfig } from 'web3-core'
-import { SolanaProvider } from './solana'
-import { Chain, ChainId } from './chain'
-import { ProviderSubscriber } from './provider-event-bus'
 
-export type RawProvider = EthereumProvider | SolanaProvider
+import { ChainTypes, Providers } from '@/enums'
+
+import { Chain, ChainId } from './chain'
+import { EthereumProvider } from './ethereum'
+import { NearProviderType } from './near'
+import { ProviderSubscriber } from './provider-event-bus'
+import { SolanaProvider } from './solana'
+
+export type RawProvider = EthereumProvider | SolanaProvider | NearProviderType
 
 export type ProviderInstance = {
   name: Providers
-  instance: RawProvider
+  instance?: RawProvider
 }
 
 export type TxRequestBody =
@@ -31,9 +34,12 @@ export type EthereumTransaction = TransactionConfig
 
 export type SolanaTransactionResponse = TransactionSignature
 
+export type NearTransactionResponse = nearProviders.FinalExecutionOutcome
+
 export type TransactionResponse =
   | EthTransactionResponse
   | SolanaTransactionResponse
+  | NearTransactionResponse
 
 export interface ProviderProxyConstructor {
   new (provider: RawProvider): ProviderProxy