+++ CHANGELOG.md
@@ -4,12 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
+
+## [1.3.0] - 2023-03-03
+### Added
+- `all`: Build to CommonJS and ES modules
+- `all`: `tsc-alias` package to use aliases in TypeScript
+- `root`: `yarn rsc` Release Sanity Check script
+- `root`: Publish to NPM GitHub Action
+- `@rarimo/nft-checkout`: Mainnet contract to the chain list
 
 ### Fixed
 - `@rarimo/provider`: Solflare wallet naming [issue](https://github.com/rarimo/js-sdk/issues/12)
+- `@rarimo/nft-checkout`: Estimate swap for input native token
 
-## 1.2.0 - 2023-02-28
+## [1.2.0] - 2023-02-28
 ### Added
 - `root`: Yarn TypeScript plugin which automatically adds @types/ packages into your dependencies when you add a package that doesn't include its own types
 - `@rarimo/provider`: Support of Phantom wallet
@@ -20,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `@rarimo/nft-checkout`: `supportedTokens` method to `INFTCheckoutOperation` interface to get supported tokens for swap on the selected chain
 - `@rarimo/nft-checkout`: PancakeSwap integration
 - `@rarimo/nft-checkout`: Ability to provide slippage in target object
+- `@rarimo/provider`: Support of Near wallet
 
 ### Changed
 - `root`: Added some more info about SDK to the readme
@@ -33,11 +43,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `root`: `@babel/*` from dependencies to simplify and speed up build process
 - `all`: `@/ -> ./src` aliases to remove `@babel/*` and all related packages
 - `@rarimo/nft-checkout`: Swap contract addresses removed from EVM mainnet chains till they will be deployed
+- `@rarimo/provider`: dependence on the web3
 
 ### Fixed
 - `@rarimo/nft-checkout`: Swap for input native token
 
-## 1.1.0 - 2023-02-20
+## [1.1.0] - 2023-02-20
 ### Added
  - `@rarimo/nft-checkout`: `createCheckoutOperation` function simplifies create NFT checkout operation.
  - `@rarimo/provider` -`createProvider` function, to simplify interface, which will create provider instance, user need to pass provider proxy implementation for needed wallet and web3 instance optionally.
@@ -49,11 +60,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `@rarimo/nft-checkout`: Approving non swap contract address, which cause error "transfer amount exceeds spender allowance"</li>
 
-## 1.0.1 - 2023-02-16
+## [1.0.1] - 2023-02-16
 ### Added
 - `@rarimo/nft-checkout`: `events` dependency to resolve a Node emulation issue for use in packagers such as Vite</li>
 
-## 1.0.0
+## [1.0.0] - 2023-02-16
 ### Under the hood changes
 - Initiated repo with `@rarimo/provider` and `@rarimo/nft-checkout` packages
 
+[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.3.0...HEAD
+[1.3.0]: https://github.com/rarimo/js-sdk/compare/1.2.0...1.3.0
+[1.2.0]: https://github.com/rarimo/js-sdk/compare/1.1.0...1.2.0
+[1.1.0]: https://github.com/rarimo/js-sdk/compare/1.0.1...1.1.0
+[1.0.1]: https://github.com/rarimo/js-sdk/compare/1.0.0...1.0.1
+[1.0.0]: https://github.com/rarimo/js-sdk/releases/tag/1.0.0